split up the BuildAstVisitor into several files
